backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
qdrant-client==1.7.0
numpy==1.24.3
torch==2.1.2
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.37.2
python-jose==3.3.0
psycopg2-binary==2.9.9
office365-rest-python-client==2.4.1
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0 